vue refactoring
