+++ .gitignore
@@ -1,3 +1,4 @@
 data
 dist
 .idea
+node_modules
