+++ README.md
@@ -1,7 +1,3 @@
-# Disclaimer
-
-I no longer have time to work on this project, I will look at pull requests if people want to update it, otherwise I can't promise it to work on the latest patch.
-
 # d2s-editor
 
 [Example](http://d2s.dschu012.dev/)
@@ -18,7 +14,8 @@ This is a basic Diablo 2 save editor for D2 and D2R. The parser uses the TXT dat
 ### How to use on your own mod
 
 1. Clone this repo or download a zip.
-2. Extract your MPQ data to a data folder in the same directory as the repo. (the tool needs TXTs, strings, pallettes, and item dc6s to work properly)
-3. Run a simple http server hosting this directory (e.g. [http-server](https://www.npmjs.com/package/http-server), [nginx](https://www.nginx.com/), [IIS](https://www.iis.net/))
-4. Go to http://localhost:${PORT}/constants.html. Follow instructions to generate the TXT constant data for the save parser.
-5. Go to http://localhost:${PORT}/ verify everything works.
+2. Extract your MPQ data to a public/data folder in the same directory as the repo. (the tool needs TXTs, strings, pallettes, and item dc6s to work properly)
+3. Run 
+   npm install
+   npm run serve
+4. Go to http://localhost:${PORT}/ verify everything works.
