--- js/html.js
@@ -1,9 +1,0 @@
-
-export default function() {
-  let tagArgs = arguments;
-  return tagArgs[0].reduce(function (accumulator, string, index) {
-    accumulator += string;
-    if (index + 1 in tagArgs) accumulator += tagArgs[index + 1];
-    return accumulator;
-  }, '');
-}
