+++ src/components/ContextMenu.vue
@@ -1,13 +1,13 @@
-import html from '../html.js';
+<template>
+  <div class="dropdown-menu" role="menu" :class="options != null ? 'show' : ''" v-bind:style="{top: evt.clientY+'px', left: evt.clientX+'px'}">
+    <button class="dropdown-item" v-for="option in options" @click.stop="onClick($event, option)"
+        :class="[option.type === 'divider' ? 'dropdown-divider' : '']" v-html="option.text">
+    </button>
+  </div>
+</template>
 
-export default {
-    template: html`
-        <div class="dropdown-menu" role="menu" :class="options != null ? 'show' : ''" v-bind:style="{top: evt.clientY+'px', left: evt.clientX+'px'}">
-            <button class="dropdown-item" v-for="option in options" @click.stop="onClick($event, option)"
-                :class="[option.type === 'divider' ? 'dropdown-divider' : '']" v-html="option.text">
-            </button>
-        </div>
-    `,
+<script>
+  export default {
     data: function(){
         return {
             options: null,
@@ -34,4 +34,5 @@ export default {
             this.options = null;
         }
     }
-}
+  };  
+</script>
