+++ src/components/inventory/ItemEditor.vue
@@ -1,10 +1,4 @@
-import Item from './Item.js';
-import ItemStatsEditor from './ItemStatsEditor.js';
-
-import html from '../../html.js';
-
-export default {
-  template: html`
+<template>
   <div>
     <div class="form-row">
       <div class="col-md-12">
@@ -144,54 +138,61 @@ export default {
         </div>
       </div>
     </span>
-</div>
-`,
-name: 'ItemEditor',
-props: {
-  id: String,
-  item: Object,
-  location: Object,
-},
-components: {
-  Item,
-  ItemStatsEditor,
-},
-data() {
-  return {
-    rarities: [{ key: 1, value: 'Low' }, { key: 2, value: 'Normal' }, { key: 3, value: 'Superior' }, { key: 4, value: 'Magic' }, { key: 5, value: 'Set' }, { key: 6, value: 'Rare' }, { key: 7, value: 'Unique' }, { key: 8, value: 'Crafted' }],
-    locations: [{ key: 0, value: 'Stored' }, { key: 1, value: 'Equipped' }, { key: 4, value: 'Cursor' }],
-    equipped_locations: [{ key: 1, value: 'Head' }, { key: 2, value: 'Neck' }, { key: 3, value: 'Torso' }, { key: 4, value: 'Right Hand' }, { key: 5, value: 'Left Hand' }, { key: 6, value: 'Right Finger' }, { key: 7, value: 'Left Finger' }, { key: 8, value: 'Waist' }, { key: 9, value: 'Boots' }, { key: 10, value: 'Gloves' }, { key: 11, value: 'Alternate Right Hand' }, { key: 12, value: 'Alternate Left Hand' }],
-    storage_pages: [{ key: 1, value: 'Inventory' }, { key: 4, value: 'Cube' }, { key: 5, value: 'Stash' }],
-    prefixes: window.constants.constants.magic_prefixes.map((e,i)=> { return { i:i, v:e }}).filter(e => e.v != null && e.v.n != null),
-    suffixes: window.constants.constants.magic_suffixes.map((e,i)=> { return { i:i, v:e }}).filter(e => e.v != null && e.v.n != null),
-    rare_names: window.constants.constants.rare_names.map((e,i)=> { return { i:i, v:e }}).filter(e => e.v != null && e.v.n != null),
-    unq_items: window.constants.constants.unq_items.map((e,i)=> { return { i:i, v:e }}).filter(e => e.v != null && e.v.n != null),
-    set_items: window.constants.constants.set_items.map((e,i)=> { return { i:i, v:e }}).filter(e => e.v != null && e.v.n != null),
-    armor_items: Object.entries(window.constants.constants.armor_items).filter(e => e[1].n != null),
-    weapon_items: Object.entries(window.constants.constants.weapon_items).filter(e => e[1].n != null),
-    other_items: Object.entries(window.constants.constants.other_items).filter(e => e[1].n != null),
-  };
-},
-methods: {
-  onEvent(type) {
-    this.$emit('item-event', { item: this.item, type: type });
-  },
-  onChildEvent(e) {
-    this.$emit('item-event', { item: e.item, type: e.type });
-  },
-  onMove() {
-    this.$emit('item-event', { item: this.item, location: this.location, type: 'move' });
-  },
-  basesByType(type) {
-    if (this.item.type_id == 3) {
-      return this.weapon_items.filter(e => e[1].nc == type || e[1].exc == type || e[1].elc == type)
-    } else if (this.item.type_id == 1) {
-      return this.armor_items.filter(e => e[1].nc == type || e[1].exc == type || e[1].elc == type)
-    } else if (this.item.type_id == 4) {
-      return this.other_items
-    } else {
-      return []
+  </div>
+</template>
+
+<script>
+  import Item from './Item.vue';
+  import ItemStatsEditor from './ItemStatsEditor.vue';
+
+  export default {
+    name: 'ItemEditor',
+    props: {
+      id: String,
+      item: Object,
+      location: Object,
+    },
+    components: {
+      Item,
+      ItemStatsEditor,
+    },
+    data() {
+      return {
+        rarities: [{ key: 1, value: 'Low' }, { key: 2, value: 'Normal' }, { key: 3, value: 'Superior' }, { key: 4, value: 'Magic' }, { key: 5, value: 'Set' }, { key: 6, value: 'Rare' }, { key: 7, value: 'Unique' }, { key: 8, value: 'Crafted' }],
+        locations: [{ key: 0, value: 'Stored' }, { key: 1, value: 'Equipped' }, { key: 4, value: 'Cursor' }],
+        equipped_locations: [{ key: 1, value: 'Head' }, { key: 2, value: 'Neck' }, { key: 3, value: 'Torso' }, { key: 4, value: 'Right Hand' }, { key: 5, value: 'Left Hand' }, { key: 6, value: 'Right Finger' }, { key: 7, value: 'Left Finger' }, { key: 8, value: 'Waist' }, { key: 9, value: 'Boots' }, { key: 10, value: 'Gloves' }, { key: 11, value: 'Alternate Right Hand' }, { key: 12, value: 'Alternate Left Hand' }],
+        storage_pages: [{ key: 1, value: 'Inventory' }, { key: 4, value: 'Cube' }, { key: 5, value: 'Stash' }],
+        prefixes: window.constants.constants.magic_prefixes.map((e,i)=> { return { i:i, v:e }}).filter(e => e.v != null && e.v.n != null),
+        suffixes: window.constants.constants.magic_suffixes.map((e,i)=> { return { i:i, v:e }}).filter(e => e.v != null && e.v.n != null),
+        rare_names: window.constants.constants.rare_names.map((e,i)=> { return { i:i, v:e }}).filter(e => e.v != null && e.v.n != null),
+        unq_items: window.constants.constants.unq_items.map((e,i)=> { return { i:i, v:e }}).filter(e => e.v != null && e.v.n != null),
+        set_items: window.constants.constants.set_items.map((e,i)=> { return { i:i, v:e }}).filter(e => e.v != null && e.v.n != null),
+        armor_items: Object.entries(window.constants.constants.armor_items).filter(e => e[1].n != null),
+        weapon_items: Object.entries(window.constants.constants.weapon_items).filter(e => e[1].n != null),
+        other_items: Object.entries(window.constants.constants.other_items).filter(e => e[1].n != null),
+      };
+    },
+    methods: {
+      onEvent(type) {
+        this.$emit('item-event', { item: this.item, type: type });
+      },
+      onChildEvent(e) {
+        this.$emit('item-event', { item: e.item, type: e.type });
+      },
+      onMove() {
+        this.$emit('item-event', { item: this.item, location: this.location, type: 'move' });
+      },
+      basesByType(type) {
+        if (this.item.type_id == 3) {
+          return this.weapon_items.filter(e => e[1].nc == type || e[1].exc == type || e[1].elc == type)
+        } else if (this.item.type_id == 1) {
+          return this.armor_items.filter(e => e[1].nc == type || e[1].exc == type || e[1].elc == type)
+        } else if (this.item.type_id == 4) {
+          return this.other_items
+        } else {
+          return []
+        }
+      }
     }
-  }
-}
-};
+  };  
+</script>
