+++ src/d2/CharPack.js
