+++ src/d2/ItemPack.js
